Start new remocon, enable sass and css modules
