+++ package.json
@@ -97,6 +97,8 @@
     "relay-compiler": "^10.1.0",
     "relay-compiler-language-typescript": "^14.0.0",
     "relay-config": "^10.1.0",
+    "sass": "^1.52.2",
+    "sass-loader": "^13.0.0",
     "ts-loader": "^8.0.11",
     "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
