+++ src/common/hooks/useQueue.ts
@@ -1,6 +1,5 @@
 import { useEffect, useState } from "react";
 import { fetchQuery, graphql, requestSubscription } from "react-relay";
-import Song from "../../remocon/Song";
 
 import environment from "../graphqlEnvironment";
 import { useQueueQueueQuery } from "./__generated__/useQueueQueueQuery.graphql";
