+++ src/remocon/components/SearchMethodsGrid/SearchMethodsGrid.module.scss
@@ -1,0 +1,3 @@
+.grid {
+  display: flex;
+}
