+++ src/remocon/components/SearchMethodsGrid/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./SearchMethodsGrid";
