+++ src/remocon/pages/HomePage.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { RouteComponentProps } from "react-router-dom";
+
+import SearchMethodsGrid from "../components/SearchMethodsGrid";
+
+const HomePage = (props: RouteComponentProps) => (
+  <>
+    <SearchMethodsGrid />
+  </>
+);
+
+export default HomePage;
