+++ src/remoconOld/NicoInfo.tsx
