+++ src/remoconOld/YoutubeInfo.css
@@ -47,5 +47,5 @@
 }
 
 .add-lyrics-section {
-  margin-top: 0em
+  margin-top: 0em;
 }
