+++ src/remoconOld/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <link
+      href="https://fonts.googleapis.com/icon?family=Material+Icons"
+      rel="stylesheet"
+    />
+    <link rel="shortcut icon" href="/icon.png" type="image/png" />
+    <title>karafriends</title>
+  </head>
+  <body>
+    <div id="root"></div>
+  </body>
+</html>
